feat(filesystem): consolidate files
It was getting kind of confusing to have code in both the 'src' and 'electron' folder, especially
now that we added the separate 'client' module. To make this a little easier, I consolidated
everything under 'src'. The subfolders 'renderer', 'main', and 'client' correspond to each of the
different processes. 'types' is another top-level subfolder
