+++ src/renderer/components/Greetings/Greetings.spec.tsx
