+++ src/renderer/components/Greetings/index.tsx
