+++ src/renderer/components/Greetings/styles.ts
