+++ src/renderer/sections/ChannelSelect/ProfilePicture.tsx
