+++ src/renderer/styles/GlobalStyle.ts
